Uploaded complete project
You can edit the number of spheres, their radii, reflection colors, and
positions. You can also change the position of the light, its color,
and the color of ambient light. Camera inputs are in the form of eye,
gaze, and tilt vectors, combined with a viewing angle, near and far
viewing planes, and a screen resolution .